refactor: split display code to a separate files
