--- src/main/kotlin/org/jetbrains/research/testspark/data/TestCasesUIContext.kt
@@ -1,10 +1,0 @@
-package org.jetbrains.research.testspark.data
-
-import java.awt.Color
-import javax.swing.JPanel
-
-data class TestCasesUIContext(
-    var testCasePanels: HashMap<String, JPanel> = HashMap(),
-    var testsSelected: Int = 0,
-    var defaultEditorColor: Color? = null,
-)
