+++ src/main/kotlin/org/jetbrains/research/testspark/data/UIContext.kt
@@ -6,6 +6,5 @@ import org.jetbrains.research.testspark.core.generation.llm.network.RequestManag
 data class UIContext(
     val projectContext: ProjectContext,
     val testGenerationOutput: TestGenerationData,
-    val testCasesUIContext: TestCasesUIContext,
     var requestManager: RequestManager? = null,
 )
