+++ src/main/kotlin/org/jetbrains/research/testspark/display/GeneratedTestsTabData.kt
@@ -1,0 +1,25 @@
+package org.jetbrains.research.testspark.display
+
+import com.intellij.ui.components.JBScrollPane
+import com.intellij.ui.content.Content
+import com.intellij.ui.content.ContentManager
+import org.jetbrains.research.testspark.core.data.TestCase
+import org.jetbrains.research.testspark.display.generatedTestsTab.TestCasePanelFactory
+import org.jetbrains.research.testspark.display.generatedTestsTab.TopButtonsPanelFactory
+import javax.swing.JPanel
+
+class GeneratedTestsTabData {
+    val testCasePanels: HashMap<String, JPanel> = HashMap()
+    var testsSelected: Int = 0
+    val unselectedTestCases: HashMap<Int, TestCase> = HashMap()
+    val testCasePanelFactories: ArrayList<TestCasePanelFactory> = arrayListOf()
+    var allTestCasePanel: JPanel = JPanel()
+    var scrollPane: JBScrollPane = JBScrollPane(
+        allTestCasePanel,
+        JBScrollPane.VERTICAL_SCROLLBAR_ALWAYS,
+        JBScrollPane.HORIZONTAL_SCROLLBAR_NEVER,
+    )
+    var topButtonsPanelFactory = TopButtonsPanelFactory()
+    var contentManager: ContentManager? = null
+    var content: Content? = null
+}
