+++ src/main/kotlin/org/jetbrains/research/testspark/display/generatedTestsTab/TopButtonsPanelFactory.kt
@@ -1,28 +1,16 @@
 package org.jetbrains.research.testspark.display.generatedTestsTab
 
-import com.intellij.openapi.progress.ProgressIndicator
-import com.intellij.openapi.progress.ProgressManager
-import com.intellij.openapi.progress.Task
-import com.intellij.openapi.project.Project
 import org.jetbrains.research.testspark.bundles.plugin.PluginLabelsBundle
-import org.jetbrains.research.testspark.bundles.plugin.PluginMessagesBundle
-import org.jetbrains.research.testspark.core.progress.CustomProgressIndicator
-import org.jetbrains.research.testspark.data.UIContext
-import org.jetbrains.research.testspark.display.custom.IJProgressIndicator
 import org.jetbrains.research.testspark.display.utils.IconButtonCreator
 import org.jetbrains.research.testspark.display.utils.TestSparkIcons
 import java.awt.Dimension
-import java.util.LinkedList
-import java.util.Queue
 import javax.swing.Box
 import javax.swing.BoxLayout
 import javax.swing.JButton
-import javax.swing.JCheckBox
 import javax.swing.JLabel
-import javax.swing.JOptionPane
 import javax.swing.JPanel
 
-class TopButtonsPanelFactory(private val project: Project, private val uiContext: UIContext) {
+class TopButtonsPanelFactory {
     private var runAllButton: JButton = createRunAllTestButton()
     private var selectAllButton: JButton =
         IconButtonCreator.getButton(TestSparkIcons.selectAll, PluginLabelsBundle.get("selectAllTip"))
@@ -37,7 +25,7 @@ class TopButtonsPanelFactory(private val project: Project, private val uiContext
     private val testsPassedText: String = "${PluginLabelsBundle.get("testsPassed")}: %d/%d"
     private var testsPassedLabel: JLabel = JLabel(testsPassedText)
 
-    private val testCasePanelFactories = arrayListOf<TestCasePanelFactory>()
+    private var numberOfPassedTests = 0;
 
     fun getPanel(): JPanel {
         val panel = JPanel()
@@ -53,18 +41,13 @@ class TopButtonsPanelFactory(private val project: Project, private val uiContext
         panel.add(unselectAllButton)
         panel.add(removeAllButton)
 
-        selectAllButton.addActionListener { toggleAllCheckboxes(true) }
-        unselectAllButton.addActionListener { toggleAllCheckboxes(false) }
-        removeAllButton.addActionListener { removeAllTestCases() }
-        runAllButton.addActionListener { runAllTestCases() }
-
         return panel
     }
 
     /**
      * Updates the labels.
      */
-    fun updateTopLabels() {
+    fun update(testCasePanels: HashMap<String, JPanel>, testsSelected: Int, testCasePanelFactories: ArrayList<TestCasePanelFactory>) {
         var numberOfPassedTests = 0
         for (testCasePanelFactory in testCasePanelFactories) {
             if (testCasePanelFactory.isRemoved()) continue
@@ -75,108 +58,30 @@ class TopButtonsPanelFactory(private val project: Project, private val uiContext
         }
         testsSelectedLabel.text = String.format(
             testsSelectedText,
-            uiContext.testCasesUIContext.testsSelected,
-            uiContext.testCasesUIContext.testCasePanels.size,
+            testsSelected,
+            testCasePanels.size,
         )
         testsPassedLabel.text =
             String.format(
                 testsPassedText,
                 numberOfPassedTests,
-                uiContext.testCasesUIContext.testCasePanels.size,
+                testCasePanels.size,
             )
         runAllButton.isEnabled = false
         for (testCasePanelFactory in testCasePanelFactories) {
             runAllButton.isEnabled = runAllButton.isEnabled || testCasePanelFactory.isRunEnabled()
         }
+
+        this.numberOfPassedTests = numberOfPassedTests
     }
 
-    /**
-     * Sets the array of TestCasePanelFactory objects.
-     *
-     * @param testCasePanelFactories The ArrayList containing the TestCasePanelFactory objects to be set.
-     */
-    fun setTestCasePanelFactoriesArray(testCasePanelFactories: ArrayList<TestCasePanelFactory>) {
-        this.testCasePanelFactories.addAll(testCasePanelFactories)
-    }
+    fun getRunAllButton() = runAllButton
 
-    /**
-     * Toggles check boxes so that they are either all selected or all not selected,
-     *  depending on the provided parameter.
-     *
-     *  @param selected whether the checkboxes have to be selected or not
-     */
-    private fun toggleAllCheckboxes(selected: Boolean) {
-        uiContext.testCasesUIContext.testCasePanels.forEach { (_, jPanel) ->
-            val checkBox = jPanel.getComponent(0) as JCheckBox
-            checkBox.isSelected = selected
-        }
-        uiContext.testCasesUIContext.testsSelected = if (selected) uiContext.testCasesUIContext.testCasePanels.size else 0
-    }
+    fun getSelectAllButton() = selectAllButton
 
-    /**
-     * Removes all test cases from the cache and tool window UI.
-     */
-    private fun removeAllTestCases() {
-        // Ask the user for the confirmation
-        val choice = JOptionPane.showConfirmDialog(
-            null,
-            PluginMessagesBundle.get("removeAllMessage"),
-            PluginMessagesBundle.get("confirmationTitle"),
-            JOptionPane.YES_NO_OPTION,
-            JOptionPane.QUESTION_MESSAGE,
-        )
+    fun getUnselectAllButton() = unselectAllButton
 
-        // Cancel the operation if the user did not press "Yes"
-        if (choice == JOptionPane.NO_OPTION) return
-
-        testSparkDisplayFactory.clear()
-    }
-
-    /**
-     * Executes all test cases.
-     *
-     * This method presents a caution message to the user and asks for confirmation before executing the test cases.
-     * If the user confirms, it iterates through each test case panel factory and runs the corresponding test.
-     */
-    private fun runAllTestCases() {
-        val choice = JOptionPane.showConfirmDialog(
-            null,
-            PluginMessagesBundle.get("runCautionMessage"),
-            PluginMessagesBundle.get("confirmationTitle"),
-            JOptionPane.OK_CANCEL_OPTION,
-            JOptionPane.WARNING_MESSAGE,
-        )
-
-        if (choice == JOptionPane.CANCEL_OPTION) return
-
-        runAllButton.isEnabled = false
-
-        // add each test generation task to queue
-        val tasks: Queue<(CustomProgressIndicator) -> Unit> = LinkedList()
-
-        for (testCasePanelFactory in testCasePanelFactories) {
-            testCasePanelFactory.addTask(tasks)
-        }
-        // run tasks one after each other
-        executeTasks(tasks)
-    }
-
-    private fun executeTasks(tasks: Queue<(CustomProgressIndicator) -> Unit>) {
-        val nextTask = tasks.poll()
-
-        nextTask?.let { task ->
-            ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Test execution") {
-                override fun run(indicator: ProgressIndicator) {
-                    task(IJProgressIndicator(indicator))
-                }
-
-                override fun onFinished() {
-                    super.onFinished()
-                    executeTasks(tasks)
-                }
-            })
-        }
-    }
+    fun getRemoveAllButton() = removeAllButton
 
     /**
      * Creates a JButton for running all tests.
@@ -189,9 +94,5 @@ class TopButtonsPanelFactory(private val project: Project, private val uiContext
         runTestButton.isContentAreaFilled = false
         runTestButton.isBorderPainted = true
         return runTestButton
-    }
-
-    fun clear() {
-        testCasePanelFactories.clear()
     }
 }
