+++ src/main/kotlin/org/jetbrains/research/testspark/display/utils/ReportUpdater.kt
@@ -3,6 +3,7 @@ package org.jetbrains.research.testspark.display.utils
 import com.intellij.openapi.project.Project
 import org.jetbrains.research.testspark.core.data.Report
 import org.jetbrains.research.testspark.core.data.TestCase
+import org.jetbrains.research.testspark.display.GeneratedTestsTabData
 import org.jetbrains.research.testspark.display.coverage.CoverageVisualisationTabFactory
 
 object ReportUpdater {
@@ -11,11 +12,12 @@ object ReportUpdater {
         report: Report,
         testCase: TestCase,
         coverageVisualisationTabFactory: CoverageVisualisationTabFactory,
+        generatedTestsTabData: GeneratedTestsTabData,
     ) {
         report.testCaseList.remove(testCase.id)
         report.testCaseList[testCase.id] = testCase
         report.normalized()
-        coverageVisualisationTabFactory.showCoverage(report)
+        coverageVisualisationTabFactory.show(report, generatedTestsTabData)
     }
 
     fun removeTestCase(
@@ -23,10 +25,11 @@ object ReportUpdater {
         report: Report,
         testCase: TestCase,
         coverageVisualisationTabFactory: CoverageVisualisationTabFactory,
+        generatedTestsTabData: GeneratedTestsTabData,
     ) {
         report.testCaseList.remove(testCase.id)
         report.normalized()
-        coverageVisualisationTabFactory.showCoverage(report)
+        coverageVisualisationTabFactory.show(report, generatedTestsTabData)
     }
 
     fun unselectTestCase(
@@ -35,9 +38,10 @@ object ReportUpdater {
         unselectedTestCases: HashMap<Int, TestCase>,
         testCaseId: Int,
         coverageVisualisationTabFactory: CoverageVisualisationTabFactory,
+        generatedTestsTabData: GeneratedTestsTabData,
     ) {
         unselectedTestCases[testCaseId] = report.testCaseList[testCaseId]!!
-        removeTestCase(project, report, report.testCaseList[testCaseId]!!, coverageVisualisationTabFactory)
+        removeTestCase(project, report, report.testCaseList[testCaseId]!!, coverageVisualisationTabFactory, generatedTestsTabData)
     }
 
     fun selectTestCase(
@@ -46,10 +50,11 @@ object ReportUpdater {
         unselectedTestCases: HashMap<Int, TestCase>,
         testCaseId: Int,
         coverageVisualisationTabFactory: CoverageVisualisationTabFactory,
+        generatedTestsTabData: GeneratedTestsTabData,
     ) {
         report.testCaseList[testCaseId] = unselectedTestCases[testCaseId]!!
         unselectedTestCases.remove(testCaseId)
         report.normalized()
-        coverageVisualisationTabFactory.showCoverage(report)
+        coverageVisualisationTabFactory.show(report, generatedTestsTabData)
     }
 }
