+++ src/main/kotlin/org/jetbrains/research/testspark/tools/evosuite/generation/EvoSuiteProcessManager.kt
@@ -204,6 +204,6 @@ class EvoSuiteProcessManager(
             e.printStackTrace()
         }
 
-        return UIContext(projectContext, generatedTestData, TestCasesUIContext(), StandardRequestManagerFactory(project).getRequestManager(project))
+        return UIContext(projectContext, generatedTestData, StandardRequestManagerFactory(project).getRequestManager(project))
     }
 }
