+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/LLMProcessManager.kt
@@ -196,6 +196,6 @@ class LLMProcessManager(
             generatedTestsData,
         )
 
-        return UIContext(projectContext, generatedTestsData, TestCasesUIContext(), requestManager)
+        return UIContext(projectContext, generatedTestsData, requestManager)
     }
 }
